// ambari-web/latest/src/screens/Metrics/type.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type WidgetInfo = {
    id: number;
    metrics: string;
    author: string;
    description: string;
    scope: string;
    properties: string;
    tag: any;
    widget_name: string;
    widget_type: string;
    time_created: number;
    display_name: string;
    values: string;
    cluster_name: string;
  }